Implemented a lot of changes to boundries
Boundries got a new script for control and narration clips were implemented as well as a lot of logic implemented, like removing and taking back the different boundries but not the ones that have already been used.
As well as a counter with narration as well that keep track of how many stations the user has finished.